Develop

+++ Godeps/_workspace/src/github.com/tendermint/tendermint/merkle/simple_tree_test.go
@@ -3,7 +3,6 @@ package merkle
 import (
 	. "github.com/eris-ltd/mint-client/Godeps/_workspace/src/github.com/tendermint/tendermint/common"
 	. "github.com/eris-ltd/mint-client/Godeps/_workspace/src/github.com/tendermint/tendermint/common/test"
-
 	"testing"
 )
 
@@ -15,10 +14,10 @@ func (tI testItem) Hash() []byte {
 
 func TestSimpleProof(t *testing.T) {
 
-	numItems := uint(100)
+	numItems := 100
 
 	items := make([]Hashable, numItems)
-	for i := uint(0); i < numItems; i++ {
+	for i := 0; i < numItems; i++ {
 		items[i] = testItem(RandBytes(32))
 	}
